Support db2
