--- .travis.yml
@@ -1,12 +1,0 @@
-# Use Node.js configuration.
-language: node_js
-
-node_js:
-  - 10
-  - 12
-  - node # latest
-
-os:
-  - linux
-  - osx
-  - windows
