+++ package.json
@@ -12,11 +12,16 @@
     "url": "git://github.com/ssbc/ssb-friends.git"
   },
   "dependencies": {
+    "bipf": "^1.4.0",
+    "flumecodec": "0.0.1",
     "flumeview-reduce": "^1.3.17",
     "layered-graph": "^1.1.1",
     "pull-flatmap": "0.0.1",
+    "pull-level": "^2.0.4",
     "pull-notify": "^0.1.1",
     "pull-stream": "^3.6.0",
+    "ssb-caps": "^1.1.0",
+    "ssb-db2": "^1.9.1",
     "ssb-ref": "^2.7.1"
   },
   "devDependencies": {
